Refactored nyxd client usages to make things bit more clear
